% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cols.R
\name{get_cols}
\alias{get_cols}
\title{gets the colours corresponding to the coloured category for use in ggplot}
\usage{
get_cols(categories, col_options = NULL)
}
\arguments{
\item{categories}{a vector of unique categories to be coloured}

\item{col_options}{(optional) customise the colours for the input unique categories}
}
\value{
a vector of colours named by the categories
}
\description{
gets the colours corresponding to the coloured category for use in ggplot
}
\examples{
categories = c('cat', 'dog', 'duck')
get_cols(categories)
}
