% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapes.R
\name{get_shapes}
\alias{get_shapes}
\title{gets the shapes corresponding to the shaped category for use in ggplot}
\usage{
get_shapes(categories, shape_options = NULL)
}
\arguments{
\item{categories}{a vector of unique categories to be coloured}

\item{shape_options}{(optional) customise the shapes for the input unique categories}
}
\value{
a vector of shapes named by the categories
}
\description{
gets the shapes corresponding to the shaped category for use in ggplot
}
\examples{
categories = c('cat', 'dog', 'duck')
get_shapes(categories)
}
